+++ src/Hyperttp/Utils/index.ts
@@ -1,2 +1,4 @@
 export { RequestBuilder } from "./RequestBuilder.js";
+export type { ResponseType } from "./RequestBuilder.js";
 export { appendQueryToUrl } from "./query.js";
+export type { RequestQuery } from "./query.js";