+++ src/index.ts
@@ -1,8 +1,11 @@
 export { HyperClient, RequestBuilder } from "./Hyperttp/index.js";
+export type { ResponseType, ShortcutRequestOptions } from "./Hyperttp/index.js";
 export { Request, PreparedRequest, UrlExtractor } from "./Hyperttp/index.js";
 
 export type { RequestQuery } from "./Hyperttp/Utils/query.js";
 export type { RequestHeaders, RequestBodyData, RequestConfig } from "./Hyperttp/Request.js";
+export type { HttpMethod, RestRequestOptions } from "@hyperttp/core/rest";
+export type { HyperClientOptions, HyperTransport, UniversalResponse } from "@hyperttp/types";
 
 export type * from "./Types/index.js";